==> Modeling/0_read_data.ipynb <==
# read data

import pandas as pd
import numpy as np

data = pd.read_csv('./US_Accidents_March23.csv')

import sys
sys.path.append('./Code_V2/')
import public_func as base

base.subset_year(data, 2022, './data_2022.csv')

data2022 = pd.read_csv('./data_2022.csv')

data2022.shape

data2022.Severity.value_counts(normalize = True)

data2022.isna().sum()

test = data2022.dropna()
test.shape

test.Severity.value_counts(normalize = True)

data2022.State.value_counts()

GA = data2022[data2022.State == 'GA']
GA.shape

severity_dfs = {f"df_severity_{level}": data2022[data2022['Severity'] == level] for level in data2022['Severity'].unique()}

df_severity_1 = severity_dfs['df_severity_1']
df_severity_2 = severity_dfs['df_severity_2']
df_severity_3 = severity_dfs['df_severity_3']
df_severity_4 = severity_dfs['df_severity_4']

data2022.Severity.value_counts(normalize = True)*10000

#random sample same percentage from each severity, in total we have 10K records

# Severity
# 2    9097.246336
# 3     442.321266
# 4     245.532928
# 1     214.899470


df1 = df_severity_1.sample(n = 215, random_state=20250316)
df2 = df_severity_2.sample(n = 9097, random_state=20250316)
df3 = df_severity_3.sample(n = 442, random_state=20250316)
df4 = df_severity_4.sample(n = 246, random_state=20250316)

df_list = [df1, df2, df3, df4]
df = pd.concat(df_list)

df.shape

df.head(5)

df['severity_binary'] = np.where(df['Severity'].isin([1, 2]), 0, 1)
GA['severity_binary'] = np.where(GA['Severity'].isin([1, 2]), 0, 1)

df.to_csv('./data10K.csv', index = False)

GA.to_csv('./dataGA.csv', index = False)
