==> Modeling/2_clean_data.ipynb <==
# data clean

import pandas as pd
import numpy as np

data = pd.read_csv('./dataGA.csv')

data.shape

data['DayOfWeek'] = pd.to_datetime(data['Start_Time'], format='mixed').dt.dayofweek
data['HourOfDay'] = pd.to_datetime(data['Start_Time'], format='mixed').dt.hour

data['Start_Time_clean']= pd.to_datetime(data['Start_Time'], format='mixed')
data['End_Time_clean']= pd.to_datetime(data['End_Time'], format='mixed')

data['accident_duration'] = (data['End_Time_clean'] - data['Start_Time_clean']).dt.total_seconds()

columns = ['Source', 'Start_Time', 'End_Time', 'End_Lat', 'End_Lng', 'Country', 'Wind_Direction', 'Severity',
           'Turning_Loop', 'Weather_Timestamp', 'Start_Time_clean', 'End_Time_clean', 'State', 'Timezone', 'Airport_Code', 'Wind_Chill(F)',
          'Street', 'City', 'County', 'Zipcode']

data = data.drop(columns = columns)

data.shape

pd.set_option('display.max_columns', None)

data.head(5)

data.shape

data.dtypes

data.Weather_Condition.value_counts()

data['Weather_Condition'].isna().sum()

#find most common used word in weather column
from collections import Counter
data['Weather_Condition'] = data['Weather_Condition'].fillna('').astype(str)
all_words = ' '.join(data['Weather_Condition']).lower().split()
word_counts = Counter(all_words)
print("Most common words:", word_counts.most_common())

def clean_weather(weather):
    weather = str(weather).lower()

    if 'fair' in weather:
        return 'fair'
    elif 'thunder' in weather or 'storm' in weather:
        return 'storm'
    elif 'rain' in weather or 'drizzle' in weather or 'shower' in weather:
        return 'rain'
    elif 'snow' in weather:
        return 'snow'
    elif 'windy' in weather:
        return 'windy'
    elif 'cloudy' in weather:
        return 'cloudy'
    elif 'fog' in weather or 'haze' in weather:
        return 'fog'
    elif weather == '':
        return 'unknown'
    else:
        return 'other' 

data['cleaned_weather'] = data['Weather_Condition'].apply(clean_weather)
data = data.drop(columns = ['Weather_Condition'])

data['cleaned_weather'].value_counts()

data.shape

cat_cols = data.select_dtypes(include=['object']).columns

# create a summary df of each categorical feature
result = pd.DataFrame({
    'Categorical_Column': cat_cols,
    'Unique_Values_Count': [data[col].nunique() for col in cat_cols]
})
print(result)

#impute missing with mode
for col in cat_cols:
    data[col].fillna(data[col].mode()[0], inplace=True)

data.isna().sum()

#if value count less then 10 then convert into dummies, greater than 10 convert into numeric
dummies = list(result[result.Unique_Values_Count <=10]['Categorical_Column'])
data = pd.get_dummies(data, columns=dummies, drop_first=True)

data.shape

import sys
sys.path.append('./Code_V2/')
import public_func as base

new_data = base.convert_numerical(data, exclusive = ['Description', 'ID'])

new_data.shape

impute missing of numeric features

num_cols = new_data.select_dtypes(include=['number']).columns
new_data[num_cols] = new_data[num_cols].apply(lambda x: x.fillna(x.median()))

new_data.isna().sum()

new_data.to_csv('./cleaned_data.csv', index=False)
