==> Modeling/3_text_mining.ipynb <==
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import TruncatedSVD

def build_TF_IDF_X(data, svd_components=150, X_addition=None):
    vectorizer = TfidfVectorizer(stop_words='english', max_df=0.5, min_df=5, norm=None) #, max_features=3000
    tfidf_matrix = vectorizer.fit_transform(data['Description'].str.lower())
    print(f"TfidfVectorizer size: {tfidf_matrix.shape}")

    feature_names = vectorizer.get_feature_names_out()
    # Compute the average TF-IDF score for each word
    avg_tfidf_scores = tfidf_matrix.mean(axis=0).A1  
    keywords_df = pd.DataFrame({'Word': feature_names, 'TF-IDF Score': avg_tfidf_scores})
    top_keywords = keywords_df.sort_values(by='TF-IDF Score', ascending=False)

    # Apply TruncatedSVD to reduce dimensionality
    svd = TruncatedSVD(n_components=svd_components)  # Adjust n_components as needed
    X_text_svd = svd.fit_transform(tfidf_matrix)

    # Convert reduced text data to DataFrame
    # X_text_reduced = pd.DataFrame(X_text_svd, columns=[f'V_{i}' for i in range(X_text_svd.shape[1])])
    X_text_reduced = pd.DataFrame(X_text_svd, columns=[f'SVD{i}' for i in range(X_text_svd.shape[1])])

    # X_text = pd.DataFrame(X_text_svd.toarray(), columns=vectorizer.get_feature_names_out())

    # Combine reduced text features with other features
    if X_addition is None:
        X = X_text_reduced
    else:
        X = pd.concat([X_text_reduced, X_addition.reset_index(drop=True)], axis=1)

    return [X, top_keywords]

data = pd.read_csv('./cleaned_data.csv')

X, top_keywords = build_TF_IDF_X(data)

top_keywords.to_csv('./top_keywords.csv',index = False)

top_keywords

X

X.to_csv('./text_mining.csv',index = False)
