==> Modeling/4_feature_selection_training_num_text.ipynb <==
import pandas as pd
import numpy as np

#read cleaned data and text mining result
clean = pd.read_csv('./cleaned_data.csv')
text = pd.read_csv('./text_mining.csv')

clean.shape, text.shape

# scale = list(set(clean.columns)-set(['ID', 'Description']))
# len(scale)

# from sklearn.preprocessing import MinMaxScaler
# scaler = MinMaxScaler()
# clean[scale] = scaler.fit_transform(clean[scale])
clean = clean[['severity_binary']]

clean.describe()

data = clean.merge(text, left_index = True, right_index=True, how = 'left')
data.shape

# data = text.copy()

# data = data.set_index('ID')
# data = data.drop(columns = ['Description'])

data.shape

### Feature selections

Boruta, Lasso, Mutual Information (Information Value), PCA

from boruta import BorutaPy

import sys
sys.path.append('./Code_V2/')
import public_func as base

output_folder = './chart/'
X = data.drop(columns = ['severity_binary'])
y = data.severity_binary

# selected_features, rejected_features = base.Boruta_selection(X, y, ff_path=output_folder + f'boruta.png')

# rejected_features

# voting_table = pd.DataFrame(X.columns, columns=['Features'])
# voting_table['Boruta'] = voting_table['Features'].isin(selected_features).astype(int)

# voting_table.Boruta.value_counts()

# from sklearn.linear_model import Lasso

# lasso = Lasso(alpha = 0.001)
# lasso.fit(X, y)
# coefficients = lasso.coef_
# print(coefficients)
# feature_coefficient = pd.DataFrame({'Feature': X.columns, 'Coefficient': coefficients})
# feature_coefficient = feature_coefficient[feature_coefficient['Coefficient'] != 0]
# feature_coefficient = feature_coefficient.sort_values(by='Coefficient', ascending=False)
# feature_coefficient.shape

# voting_table['Lasso'] = voting_table['Features'].isin(feature_coefficient.Feature).astype(int)

# voting_table.Lasso.value_counts()

# from sklearn.feature_selection import mutual_info_classif

# X_columns = X.columns
# mi_features = base.mutual_information_feature_selection(X, y, alpha = 0.001, feature_names = X_columns, ff_path = output_folder+'mi.png')

# len(mi_features)

# voting_table['MI'] = voting_table['Features'].isin(mi_features).astype(int)

# voting_table.MI.value_counts()

# from sklearn.decomposition import PCA

# pca = PCA(n_components=2)
# pca.fit(X)

# explained_variance = pca.explained_variance_ratio_
# print("Explained variance ratio for each component:", explained_variance)

# components = pca.components_

# pca_df = pd.DataFrame(components, columns=X.columns)
# print("PCA components (loadings):")
# print(pca_df)

# top_features = pca_df.iloc[0].abs().sort_values(ascending=False).head(15)
# print("Top 5 features for the first principal component:", top_features)

# voting_table.fillna(0)
# voting_table['total_selected'] = voting_table.Boruta + voting_table.Lasso + voting_table.MI

# voting_table.total_selected.value_counts()

# selected = list(voting_table[voting_table.total_selected >=2].Features)

# X = X[selected]

#split data
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state = 20250322)

models = base.create_models()
base.training_models(X_train, X_test, y_train, y_test, 'num', './chart/')

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
import lightgbm as lgb
import xgboost as xgb


models = {
    "Logistic Regression": LogisticRegression(class_weight='balanced', max_iter=1000),
    "KNN": KNeighborsClassifier(),
    "Decision Tree": DecisionTreeClassifier(),
    "Random Forest": RandomForestClassifier(),
    "LightGBM": lgb.LGBMClassifier(),
    "XGBoost": xgb.XGBClassifier()}

results = {}

for model_name, model in models.items():
    model.fit(X_train, y_train)
    y_pred = model.predict(X_test)
    
    accuracy = accuracy_score(y_test, y_pred)
    precision = precision_score(y_test, y_pred)
    recall = recall_score(y_test, y_pred)
    f1 = f1_score(y_test, y_pred)
    
    results[model_name] = {
        "Accuracy": accuracy,
        "Precision": precision,
        "Recall": recall,
        "F1 Score": f1}

import pandas as pd
results_df = pd.DataFrame(results).T

print(results_df)
